detector_app/requirements.txt
flask
grpcio
grpcio-tools
futures
prometheus_client
prometheus_flask_exporter